% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawtangle.R
\name{drawtangle}
\alias{drawtangle}
\title{Making tanglegrams of microbiome tree matched with snp trees}
\usage{
drawtangle(snptree, micro.dend, cordata.best, chr.num)
}
\arguments{
\item{snptree}{A list of dendrograms generated from snp data}

\item{micro.dend}{A dendrogram generated from microbiome relative abundance data}

\item{chr.num}{The chromosome number from which the snps are located}

\item{cordata}{A dataframe which includes positions and corrlation values of highly correlated snp trees}
}
\value{
Plots highlt correlated tanglegrams with SNP location on chromosome in basepairs
}
\description{
Making tanglegrams of microbiome tree matched with snp trees
}
\examples{
set.seed(1234) #simulate microbiome relative abundance dataframe
x <- rnorm(n=120, mean=5, sd=20)
abundance=matrix(x,10,12)
rname=paste("bacteria",1:10,sep="")
cname=paste("species", 1:12, sep="")
rownames(abundance)=rname
colnames(abundance)=cname

mymicro.dend=microtree(microdata=abundance, nboot=10) #cluster strains by microbiome

chrX = SampSnp()  #load SNP data
strain.names=colnames(abundance) #assign the same strain names
mysnptree=maketree(snpdata=chrX, strain.names=strain.names) #cluster strains by every 100 SNPs

mycordata=snpcor(snpdata=chrX, snptree=mysnptree, micro.dend=mymicro.dend) #calculate correlation coefficient for micribiome dendrogram and each SNP dendrogram
mycordata.best=snpcor.best(cordata=mycordata,threshold = 0.95) #find highly correlated SNP dendrograms
drawtangle(snptree=mysnptree, micro.dend=mymicro.dend, cordata.best=mycordata.best, chr.num="X") #plot out as tanglegrams

}
\author{
Xingyao Chen
}

