% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finddistance.R
\name{finddistance}
\alias{finddistance}
\title{Clustering strains based on dissimilarity of SNPs}
\usage{
finddistance(snpdata, start = 1, strain.names)
}
\arguments{
\item{snpdata}{SNP data downloaded from opensource databases such as <http://msub.csbio.unc.edu/>or <http://www.informatics.jax.org/javawi2/servlet/WIFetch?page=snpQF>}

\item{start}{starting row of SNP, default is 1}

\item{strain.names}{a character vector of matching strain names}
}
\value{
an object of class dendrogram
}
\description{
One drendrogram for every 100 snps
}
\author{
Xingyao Chen
}

