% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maketree.R
\name{maketree}
\alias{maketree}
\title{Generating dendrograms for the entire chromosome}
\usage{
maketree(snpdata, strain.names)
}
\arguments{
\item{snpdata}{SNP data downloaded from opensource databases such as <http://msub.csbio.unc.edu/>or <http://www.informatics.jax.org/javawi2/servlet/WIFetch?page=snpQF>}

\item{strain.names}{a character vector of strain names}
}
\value{
a large list of dendrograms
}
\description{
Uses the {finddistnace} function to cluster strains based on dissimilarity of SNPs. One drendrogram for every 100 snps
}
\examples{
chrX = SampSnp() #load SNP data
strain.names=c(paste("strain", 1:(ncol(chrX)-1), sep="")) #assign strain names, strain names must match between microbiome and SNP dendrograms
mysnptree=maketree(snpdata=chrX, strain.names=strain.names)
mysnptree
plot(mysnptree[1])

}
\author{
Xingyao Chen
}

