% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microtree.R
\name{microtree}
\alias{microtree}
\title{Clustering strains base on microbiome relative abundance data}
\usage{
microtree(mircodata, nboot = 1000)
}
\arguments{
\item{nboot}{The number of bootstrap replications. The default is 1000.}

\item{microdata}{A matrix of relative abundance data}
}
\value{
An object of dendrgram class
}
\description{
Calculates p-values for hierarchical clustering via multiscale bootstrap resampling.
}
\details{
Hierarchical clustering is done for given data and p-values are computed for each of the clusters.
}
\examples{
set.seed(1234)  #simulate microbiome relative abundance data
x <- rnorm(n=120, mean=5, sd=20)
abundance=matrix(x,10,12)
rname=paste("bacteria",1:10,sep="")
cname=paste("species", 1:12, sep="")
rownames(abundance)=rname
colnames(abundance)=cname
#
microdend=microtree(microdata=abundance, nboot=10)
microdend
plot(microdend)

}
\author{
Xingyao Chen
}

