% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampPheno.R
\name{SampPheno}
\alias{SampPheno}
\title{Loading a sample phenotype data}
\usage{
SampPheno()
}
\value{
a large dataframe of SNP data from various mouse strains
}
\description{
dataset originally downloaded from <http://phenome.jax.org/>
}
\examples{
chrX=SampSNP()
head(chrX)

}
\author{
Xingyao Chen
}

