% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SampSNP.R
\name{SampSNP}
\alias{SampSNP}
\title{Loading a sample SNP data}
\usage{
SampSNP()
}
\value{
a large dataframe of SNP data from various mouse strains
}
\description{
dataset originally downloaded from <http://msub.csbio.unc.edu/>
}
\examples{
chrX=SampSnp()
head(chrX)

}
\author{
Xingyao Chen
}

