% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawtangle.R
\name{drawtangle}
\alias{drawtangle}
\title{Matching dendrograms}
\usage{
drawtangle(snptree, micro.dend, cordata.best, chr.num)
}
\arguments{
\item{snptree}{A list of dendrograms generated from snp data}

\item{micro.dend}{A dendrogram generated from quantitative trait data}

\item{chr.num}{The chromosome number from which the snps are located}

\item{cordata}{A dataframe which includes positions and corrlation values of highly correlated snp trees}
}
\value{
Plots highlt correlated tanglegrams with SNP location on chromosome in basepairs
}
\description{
Making tanglegrams of trait dendrograms matched up with snp dendrograms
}
\examples{
set.seed(1234) #simulate quantitatie trait dataframe
x <- rnorm(n=(16*12), mean=10, sd=10)
trait=matrix(x,16,12)
rname=paste("trait",1:16,sep="")
cname=paste("strain", 1:12, sep="")
rownames(trait)=rname
colnames(trait)=cname

mymicro.dend=microtree(microdata=trait, nboot=10) #cluster strains by microbiome

chrX = SampSNP()  #load SNP data
strain.names=colnames(trait) #assign the same strain names
mysnptree=maketree(snpdata=chrX, strain.names=strain.names) #cluster strains by every 100 SNPs

mycordata=snpcor(snpdata=chrX, snptree=mysnptree, micro.dend=mymicro.dend) #calculate correlation coefficient for micribiome dendrogram and each SNP dendrogram
mycordata.best=snpcor.best(cordata=mycordata,threshold = 0.95) #find highly correlated SNP dendrograms
drawtangle(snptree=mysnptree, micro.dend=mymicro.dend, cordata.best=mycordata.best, chr.num="X") #plot out as tanglegrams

}
\author{
Xingyao Chen
}

