% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpcor.best.R
\name{snpcor.best}
\alias{snpcor.best}
\title{Subestting for highest correlations}
\usage{
snpcor.best(cordata, threshold)
}
\arguments{
\item{cordata}{A dataframe generated from the function snpcor}

\item{threshold}{Percent thrshold (0<threshold<1) to search for highest correlations, default is 0.95}
}
\value{
a subsetted correlation dataframe
}
\description{
Subsets dataframe from {snpcor} function to contain only the highly correlated SNP dendrograms
}
\examples{
set.seed(1234) #simulate microbiome relative abundance dataframe
x <- rnorm(n=120, mean=5, sd=20)
abundance=matrix(x,10,12)
rname=paste("bacteria",1:10,sep="")
cname=paste("species", 1:12, sep="")
rownames(abundance)=rname
colnames(abundance)=cname

mymicro.dend=microtree(microdata=abundance, nboot=10) #cluster strains by microbiome

chrX = SampSNP()  #load SNP data
strain.names=colnames(abundance) #assign the same strain names
mysnptree=maketree(snpdata=chrX, strain.names=strain.names) #cluster strains by every 100 SNPs

mycordata=snpcor(snpdata=chrX, snptree=mysnptree, micro.dend=mymicro.dend) #calculate correlation coefficient for micribiome dendrogram and each SNP dendrogram
mycordata.best=snpcor.best(cordata=mycordata,threshold = 0.95) #find highly correlated SNP dendrograms
}
\author{
Xingyao Chen
}

