% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpcor.R
\name{snpcor}
\alias{snpcor}
\title{Correlating SNP dendrograms with the microbiome dendrogram
Calculate Baker's Gamma correlation coefficient between two trees}
\usage{
snpcor(snpdata, snptree, micro.dend)
}
\arguments{
\item{snpdata}{snpdata SNP data downloaded from opensource databases such as <http://msub.csbio.unc.edu/>or <http://www.informatics.jax.org/javawi2/servlet/WIFetch?page=snpQF>}

\item{snptree}{a list of dendrograms generated from snp data}

\item{micro.dend}{a dendrogram generated from microbiome relative abunance data}
}
\value{
a data frame of positions, snp chromosomal location, and correlation coefficient
}
\description{
Correlating SNP dendrograms with the microbiome dendrogram
Calculate Baker's Gamma correlation coefficient between two trees
}
\examples{
set.seed(1234) #simulate microbiome relative abundance dataframe
x <- rnorm(n=120, mean=5, sd=20)
abundance=matrix(x,10,12)
rname=paste("bacteria",1:10,sep="")
cname=paste("species", 1:12, sep="")
rownames(abundance)=rname
colnames(abundance)=cname

mymicro.dend=microtree(microdata=abundance, nboot=10) #cluster strains by microbiome

chrX = read.csv("~/Biomy/chrX.csv") #load SNP data
strain.names=colnames(abundance) #assign the same strain names
mysnptree=maketree(snpdata=chrX, strain.names=strain.names) #cluster strains by every 100 SNPs

mycordata=snpcor(snpdata=chrX, snptree=mysnptree, micro.dend=mymicro.dend) #calculate correlation coefficient for micribiome dendrogram and each SNP dendrogram

}
\author{
Xingyao Chen
}

