% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traittree.R
\name{traittree}
\alias{traittree}
\title{Clustering strains base on quantitative trait data}
\usage{
traittree(traitdata, nboot = 1000)
}
\arguments{
\item{traitdata}{A matrix of relative trait data}

\item{nboot}{The number of bootstrap replications. The default is 1000.}
}
\value{
An object of dendrgram class
}
\description{
Calculates p-values for hierarchical clustering via multiscale bootstrap resampling.
}
\details{
Hierarchical clustering is done for given data and p-values are computed for each of the clusters.
}
\examples{
set.seed(1234)  #simulate quantitative trait data
x <- rnorm(n=(16*12), mean=10, sd=30)
trait=matrix(x,16,12)
rname=paste("trait",1:16,sep="")
cname=paste("strain", 1:12, sep="")
rownames(trait)=rname
colnames(trait)=cname

traitdend=traittree(traitdata=trait, nboot=10)
traitdend
plot(traitdend)

}
\author{
Xingyao Chen
}

